Se arreglaron las pantallas que estaban con errores en la codificación
