+++ src/pages/NinosProducto/NinosProducto.css
@@ -29,6 +29,7 @@
 .info-producto-Ninos h3 {
   color: var(--primary-color-dark);
   font-size: 2rem;
+  margin-top: 30px;
 }
   
 .info-producto-Ninos h4 {
@@ -62,6 +63,8 @@
   
 .ver-mas-Ninos {
   text-decoration: underline;
+  position: relative;
+  top: 13px;
 }
   
 .carrito-Ninos {
@@ -82,6 +85,8 @@
   justify-content: flex-end;
   height: 50px;
   cursor: pointer;
+  position: relative;
+  top: -33px;
 }
   
 .ver-mas-Ninos {
@@ -89,10 +94,14 @@
 }
   
 .division-productos-Ninos {
-  margin: 0 70px;
+  margin: 0 auto; /* Margen igual en ambos lados */
   background-color: var(--primary-color-dark);
   height: 2px;
   border: none;
+  position: relative;
+  top: -40px;
+  width: calc(90% - 30px); /* Hace que sea más largo al reducir el margen total */
+  left: 2px;
 }
   
 /* ------------------ Productos -------------------- */
